% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird.observations}
\alias{bird.observations}
\title{Bird observation}
\format{
A data frame with 2019 observations and 13 variables:
\describe{
\item{Species.No}{An identifier for the species}
\item{Genus}{The genus the bird belongs to}
\item{Species}{The scientific name of the species (without the genus part)}
\item{Common.Name}{The English name of the species}
\item{CBRC.Review}{Whether the bird has been reviewed by the California Bird Records Committee}
\item{Date.start}{The start date of the observation (as character!)}
\item{Date.end}{The last observation date (if any)}
\item{Number}{The number of individuals}
\item{Location}{The location of the observation - i.e. the name of the place}
\item{County}{County}
\item{Observer1}{The first (primary) observer}
\item{Other.Obs}{Other Observer(s), comma-separated}
\item{Notes}{Notes}
}
}
\usage{
bird.observations
}
\description{
This dataset contains bird observations of the bird monitoring group called
Golden Gate Audubon Society. They are active in the the San Francisco Bay
area.
}
\examples{
head(bird.observations)
}
\references{
http://goldengateaudubon.org/birding-resources/observations/
}
\keyword{datasets}
