% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{britain.species}
\alias{britain.species}
\title{Bird species on the British isles}
\format{A data frame with 42 observations and 7 variables:
\describe{
\item{island}{The name of the island}
\item{area}{The surface area of the island, in square kilometer}
\item{elevation}{Maximum elevation in meters}
\item{soil.types}{The number of different soil types found on the island}
\item{latitude}{Midpoint of latitude range in degrees north}
\item{dis.from.britain}{Distance in kilometers}
\item{num.species}{Number of observed species}
}}
\usage{
britain.species
}
\description{
Numbers of bird species observed on 42 British Isles, including area,
latitude information, distance from Britain and soil variation.
}
\examples{
head(britain.species)
}
\references{
M.P. Johnson and D.S. Simberloff (1974). "Environmental Determinants
  of Island Species Numbers in the British Isles," Journal of Biogeography,
  Vol. 1, pp. 149-154

  E.D. McCoy and E.F. Connor (1976). "Enviromnmental Determinants of
  Island Species Number in the British Isles: A Reconsideration,"
  Journal of Biogeography, Vol. 3, pp. 381-382.
}
\keyword{datasets}
