% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_netherlands}
\alias{cancer_netherlands}
\title{Cancer incidence, prevalence, survival and mortality in the Netherlands}
\format{
A data frame with 7525 observations and 5 variables:
\describe{
\item{Kankersoort}{The type of cancer (Dutch name)}
\item{Periode}{The year}
\item{Geslacht}{Gender}
\item{Leeftijdsgroep}{Age group}
\item{Aantal}{Count}
}
}
\source{
[https://iknl.nl/nkr-cijfers](https://iknl.nl/nkr-cijfers)

A broad selection was made, split over age groups, cancer types and gender (when available) and
downloaded as Excel. The long-format data was converted to an R object.
}
\usage{
cancer_netherlands
}
\description{
Data was downloaded from "Integraal Kankercentrum Nederland" on 2020-01-27
}
\keyword{datasets}
