\name{SimulVals}
\alias{SimulVals}
\title{Pre-simulated Joint Binary Probabilities}
\description{
  This variable provides a pre-fabricated result from
  \code{\link{simul.commonprob}} such that it is normally not necessary
  to use this (time consuming) function, and is used by
  \code{\link{rmvbin}}.
}
\usage{
SimulVals
}
\references{
  \bibshow{bindata::Leisch+Weingessel+Hornik:1998}
}
\author{Friedrich Leisch}
\seealso{
  \code{\link{simul.commonprob}}, \code{\link{rmvbin}}
}
\keyword{sysdata}
